python-scripts/requirements.txt
psycopg2-binary
configparser